producer/requirements.txt
asyncpg>=0.29.0
python-dotenv>=1.0.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
